requirements.txt
streamlit
PyPDF2
python-docx
gpt4all
